+++ src/components/Hero.jsx
@@ -2,10 +2,10 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 
 
-const Hero = () => {
+export const Hero = () => {
   return (
     <div className='text-blue-600'>
-        <div className='max-w-[800px] mt-[96px] w-full h-screen mx-auto text-center flex flex-col'>
+        <div className='max-w-[800px] mt-[56px] w-full h-screen mx-auto text-center flex flex-col'>
         <p className='font-bold p-2 text-3xl uppercase'>Quick loan service</p>
         <h1 className='text-blue-950 md:text-7xl sm:text-6xl text-4xl font-bold md:py-6'>Start Creating Your Dream with Us.</h1>
         <div>
@@ -19,4 +19,3 @@ const Hero = () => {
   )
 }
 
-export default Hero
