+++ src/components/index.js
@@ -1,17 +1,12 @@
-import React from 'react'
-import Hero from './Hero'
-import Footer from './Footer'
-import Navbar from './Navbar'
 
-const index = () => {
-  return (
-    <div>
-        <Footer />
-        <Hero />
-        <Navbar />
 
-    </div>
-  )
-}
+import { Footer } from './Footer';
+import { Navbar } from './Navbar';
+import { NavItems } from './NavItems';
+import { Hero } from './Hero';
 
-export {Hero, Footer, Navbar};
+
+export * from './Footer';
+export * from './NavItems';
+export * from './Navbar';
+export * from './Hero';
