+++ src/pages/About.jsx
@@ -1,9 +1,8 @@
 import React from 'react'
 
-const About = () => {
+export const About = () => {
   return (
     <div>About</div>
   )
 }
 
-export default About
