+++ src/pages/Contact.jsx
@@ -1,9 +1,8 @@
 import React from 'react'
 
-const Contact = () => {
+export const Contact = () => {
   return (
     <div>Contact</div>
   )
 }
 
-export default Contact
