+++ src/pages/Login.jsx
@@ -1,9 +1,8 @@
 import React from 'react'
 
-const Login = () => {
+export const Login = () => {
   return (
     <div>Login</div>
   )
 }
 
-export default Login
